conditionally rendering story component
