+++ frontend/src/components/Story/index.js
@@ -1,23 +1,26 @@
 import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
 
 import { getAllStories } from '../../store/story'
 
 function Story() {
     const dispatch = useDispatch();
-    const stories = useSelector(state => state.story)
-    // console.log(typeof stories); =obj
-    useEffect(() => {
-        dispatch(getAllStories())
-    }, [dispatch])
+    const { storyId } = useParams();
+    const story = useSelector(state => state.story[storyId])
+    // console.log(story)
+    // useEffect(() => {
+    //     dispatch(getAllStories())
+    // }, [dispatch])
 
     return (
-        Object.values(stories).map(story => (
-            <div key={story.id} className='story-container'>
-                <h2 className='story-title'>{story.title}</h2>
-                <p className='story-body'>{story.body}</p>
-            </div>
-        ))
+        story ?
+        <div key={story.id} className='story-container'>
+            <h2 className='story-title'>{story.title}</h2>
+            <p className='story-body'>{story.body}</p>
+        </div>
+        :
+        <div>...Loading</div>
     )
 }
